Backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
celery>=5.3
celery[redis]>=5.3
redis>=5.0
psycopg2-binary>=2.9
requests>=2.31
python-dotenv>=1.0
django-cors-headers>=4.3
dj-database-url>=2.1
